requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.2.3
requests>=2.28.0
pytest==8.2.2
httpx==0.27.2